divrep: (experimental) order testers by most knowledge
This change makes the common cases more informative but (probably)
slower - we check for gcd(k, n) and tau(gcd(k, n)), then choose the
order of testing k by descending tau, then descending gcd, then
ascending k.

So given a search for f(18, 8), we'll search in the order
6, 3, 2, 4, 8, 1, 5, 7. In practice, this seems to mean we put the
tests most likely to pass first, so the (301) result lines are more
informative about the rate of passes; but that means we reject
slower, so factorize more on average per number tested.

See the TODO for an alternative approach.