divrep: bugfix on fix_square
This could wrongly choose to fix_square for some composite x: it was
supposed to assert prior constraints such that gcd(x, d)=1, but actually
checked only that gcd(x, d) != x. This went wrong for n=108, claiming that
x=6 fixed a square for k >= 7, and therefore no solution for k >= 7 was
possible. Only when a happy accident showed a solution for k=7 did this
show up: similar cases will need to be re-checked.