divrep: extra handling for n = p^5
Add support via new subclass Constraint::Fact for n = p^5, which admits
the special factorization n + pd = pq^2 => n + xd = (aq - bp)(aq + bp)
for some x, a, b.

For now we use this only if the new '-fact' option is supplied to gtauseq.

Additional logic that makes this irrelevant for p=3 documented in
Constraint::Fact, and added as an exception in gtauseq.

You can supply the new '-ix' option to gtauseq to ignore exceptions:
they are intended largely to short-circuit harness calculations.