gtauseq: allow any digit for -mm
The '-mm25=1/0' form specifies a '-m25=1' constraint for d, and one of
the possible corresponding forms as a constraint for y after a fix_power
notes that d = (xy^z - n) / k. The final digit was allowed to be 0 or 1,
originally intended to pick one of the two values possible when we only
supported z = 2; we now allow any digit (but we'd need to express this
format differently if we wanted to support it for z > 10).