divrep: avoid crash when final value is non-integer
When there is a fixed power (so we are using Constraint::Power), the
"current value" is actually derived from the underlying power. We take
care to ensure the derived value is always an integer - except when
we have exceeded the -x limit, at which point our attempt to show
the final "301" line can crash with a "not divisible by k" error.

We use action-at-a-distance to say, for that special case, that we
can live with a result that is not an exact integer; in that case
we show the floor() of the result appended with ":LOOSE".