divrep: suggest how to shard the work
New program find-shard suggests what shards to split the work into for
finding f(n, f). Invoked with 5 mandatory options n, f, d, optc, shards,
where d is the greatest difference already checked to, optc is the number
of moduli to check, and shards is the maximum number of shards to suggest.

If called a second time with the same arguments except for increased
shards, it guarantees that the suggested shard modulus will be a multiple
of that previously suggested. (This guarantee does not hold if d or optc
changes, though it is likely to be stable if optc changes only a small
amount, or d only by a small multiple, which is what we'd usually expect.)

Output is a single line giving the sharding modulus, and the values
permitted modulo that modulus. There may be fewer values than the
number of shards requested if no better option was available.