divrep: rewrite pack_sc, _potency, frequency
pack_sc() can now deliver a slightly more optimal selection of moduli;
in practice the benefit seems to be 0.5%-1% reduction in the number of
modular tests required to rule out the same proportion of values.

_potency() and thus frequency() are now a more accurate reflection of
the proportion of numbers in a range that will pass through all modular
tests and thus need factorization checks. frequency() is not expected
to be perfect, since the way one testable modulus interacts with another
varies in ways we cannot capture.

Entries in the "303" diagnostic line are expanded from "$values/$mod"
to "$values/$possible:$mod" for each modulus, to show that the actual
benefit from testing modulo $mod is to reduce the search space
by a factor $values/$possible.