divrep/harness: reduce range for optimizing runs
By re-parsing the last run, we can check by what factor the number of
factorizations required will increase for an optimizing (-ta) run,
and scale down the range we request accordingly.

We also take advantage of the opportunity to move the remaining fragment
of logic used to determine whether to do an optimizing run from Run to
TauF, where it belongs.