divrep/harness: don't recompute useless test order
For (n, k) with very high frequency, we may find that an optimizing run
sees no candidates to test by factorization. In that case we ended up
not setting a test order, as a result of which we continued to attempt
optimizing runs every time (eg in the case of (2401, 3) before the
exception was added in commit af62cd5549).

Instead, set a simple [1 .. k] test order in that case to avoid further
attempts to recompute: in this case so few factorizations are needed
that the order doesn't actually matter.