divrep: fix off-by-one bug for pending constraints
When adding constraints, we supply also the value above which it should
be applied; if that is greater than our starting value, we initially
place it on a "pending" stack, to be reconsidered if and when calculation
gets that far.

Recovering constraints from the pending stack was handled accurately, but
the initial choice whether to apply the constraint immediately or stack it
had an off-by-one error.

As a result, every run could potentially wrongly reject the very first
value to be checked (ie $optn) if by chance a constraint should have
applied precisely from $optn + 1.

To correct for this, we additionally scanned all logs of prior runs, and
by parsing out the "300" (init) lines discovered optn, optx, n and k for
each of them. Each was then re-run with -n[$optn] -x[$optn] -c1 $n $k,
and the results verified as terminating with a 200 (success), 500 (fail),
402 (all values disallowed) or 201 (dependent) line; for those that
emitted 200, this was then cross-checked against recorded results, but
happily there were no examples where we had failed to spot a result due
to this bug.

(Bug was spotted when casually looking at the just-found result 450
for addseq(18, 2) - attempting to reproduce that with
  gtauseq -ya -n450 -x450 -c100 18 2
failed, but the same invocation with "-c1" succeeded. Analysis showed
that for -c30 we were finding a constraint that disallowed values
d == 0 (mod 30) for d > 450.)