divrep: check_mult is tauseq-specific
It must move to be a type-specific method.

The possibility of dependency may exist for addseq (eg f(16,k) = 2f(8,k)
for k >= 9), but I do not yet know if and when that can be guaranteed.
Note that for k >= 9 we find f(8,k) == 1 (mod 2), whereas for all smaller
k they are even.