divrep: extend -p option for upperlim
The list of powers to be filled in may now be distinct for each element,
by giving C<-p> a semicolon-separated list to be matched up with the
C<-i> entries in order.

Thus C<-p"2,4;4" -i0,2> will aim to fill element 0 with a square and a
fourth power, and element 2 with only a fourth power.