divrep oul: fully handle f(n, 2)
For now this is inefficient and hacky, but can find all f(n, 2) for
n in (2, 4, .. 100) in 724s.

The main inefficiency is that for a factorization such as 7.7.2, we
will generate both 2^6.3^6 and 3^6.2^6, potentially quadrupling the
work.

The main hack is the '.. || $x <= 3' in find_two_r, to suppress
attempting to generate squares. This was added because for f(78, 2)
we were finding candidate c=25429860528128 after 11s, but since
(c/2^12/3)^(1/2) = 45491 and (c/3^12/5)^(1/2) = 3093, we would then spin
trying some 2e6 different combinations of the two squared factors, for
most of which there will be 0 or 1 pair to test. Leaving them out, however
means we have just c/6^12 = 11682 pairs to test.

Resolving the hack means doing that calculation on the fly, and finding
some stable heuristic to decide whether to generate or not.