divrel oul: generalize find_two() to find_any()
This generalizes the existing incomplete algorithm, which already works
for some cases with k=3. However for cases where n=2(mod 4), and for
larger k, we will need additional logic: for each prime p < k, we need
to consider separately the cases where both a_i and a_{i+p} are divisible
by p.

In particular, for a case like f(6,3) we initially allocate 2^2 to a_0,
then require a_0 = 2^2.p, a_2 = 2.q^2, a contradiction. Since we have
no upper limit at this point, we loop forever trying to find one.

However this happily finds new minima for f(n,3), n in { 44, 52, 68, 76,
84, 88, 92 }. It also finds the known upper limit 1310430193807280626
for f(100,3) in 230s, but then takes a long time looking for further
improvements; we may need more refinement to complete that in a
reasonable time.