divrep oul: don't directly assign factorizations of tau
Rather than prepare all possible factorizations in advance and then
iterate over the cartesian product of assigning each possible
factorization to each element, modify recurse() to wrap the existing
loop over primes with an outer loop that sets the power factor for this
prime in turn to each of the divisors of tau that include the highest
prime factor dividing tau.

This saves us a massive amount of duplication: previously if we were
shortcircuiting with an early call to walk_v(), we would uselessly
repeat that work for each factorization of the elements being skipped.

For efficiency we prepare each possible list of power factors in advance;
we also allow for repeated factors by ensuring that when a prime factor
is repeated in tau, we will not assign them in descending order; thus
when tau = 100, we will assign [5, 5], [5, 10], [10, 10] but not [10, 5].