divrep oul: experimental new option -r<n> for ratchet
Let P be the maximum prime we can allocate at this recursion level
(and thus a proxy for how many times we will recurse from here);
and let W be the number of values we'd have to check if we called
walk_v right now. We were choosing to make the short-circuit call
to walk_v whenever W < P, but we now allow a ratchet R (integer)
and short-circuit only when WR < P.

Ideally we'd be able to calculate the ideal ratchet, and remove the
need to set it. To aid that we additionally count and log both the
total number of calls to recurse() and the number of values tested
in walk_v().

A high R seems especially valuable when calculating f(2^x p, k) as
p and k increase; for example f(44, 4) is fastest when 8 < R < 16,
but f(44, 5) benefits from R > 128.