divrep oul: prep all possible mintau
Prepare possible values of mintau for all cases we may be asked for -
with tau being any divisor of m = n/p, where p is the highest prime
dividing n; and with each possible selection of relevant primes being
suppressed.

For each tau this results in a list of bit vectors representing
suppressed primes, for which we then return the corresponding value
of the first such for which non of the relevant primes are suppressed.

It might be possible to speed this up further, first by making a 2^x
lookup for each of the possible bit vectors (at a risk of requiring
a ruinously large amount of memory when x is too high), secondly by
maintaining the current bit vector for each element rather than
calculating it each time (at a risk that the cost of maintenance
would exceed the cost of on-demand calculation). But the relative
cost of calls to mintau() now show low enough in profiling that
these steps are probably not worth pursuing.