divrep oul: bugfix: explicitly allow good multiples
Allocating any p^x was marking p as "used" and inhibiting any other
allocation of p^x; but this can cause us to miss values when 6 | n,
k > 4, eg with the pattern 2^5 . . . 2^2. Check for this case, and
do not suppress p if it was used in v_i and we are allocating in v_j
with p^2 | (i - j). (But explicitly do suppress if this would give
us the same power of 2 twice.)

Needed to recheck all results (6 | n, k > 4) that had not been
independently verified by harness ((24,8), (24,9), (48,6), (96,5)),
but all checked ok.