divrep oul: special case handling for n=4p, k>=4
For these cases, restrictions on powers of 2 are a powerful constraint.

TODO: when allocating these powers forces the unallocated portion to be
square, we immediately handle that case with a simple loop over possible
square roots; however this is useless if the range of possible values is
still too great, so we need a way to flag the case and allocate more.