divrep oul: extend scope of rec2
The approach is valid as is for n=4x even when x is not prime, as long
as it is odd. With small changes it could be valid for even x too, but
we'd want to improve the fallback "v_i is square" handling before
addressing that.

So for n <= 100, this should improve things for {36, 60, 84, 100}.