divrep oul: improved validity checking for repeated primes
When assigning p^x to v_i when we already have p^y assigned to v_j,
check for more cases that would make the combination invalid to permit
immediate rejection.

To facilitate this, we also change v[i]{pu} from storing a simple
boolean to say a given prime was "used" to storing the value of x.

Also remove the limit k=49 above which the previous version of these
checks would not have worked.