divrep oul: new default start index, and new option '-s'
By default we allocated the first prime power at index floor((k-1)/2),
(and then other indices in order from index 0), since that helped avoid
getting stuck on allocations that could never yield a solution for a
particular class of cases. However recent changes improve our ability to
detect and avoid such allocations, and demonstrate that in other cases
the default was actively worse.

Hence we change the default to allocate to index 0 first, but allow that
to be overridden by specifying -s$n, with $n in the range 0 to k-1.

When continuing a previously started run with -r, this option is ignored:
the value is discovered from the header line instead. It was necessary to
manually edit the header line of incomplete runs to reflect the start index
they had actually used.

The option is also ignored when the "rec2" optimization is used, ie when
n = 4q with q odd and != 0 (mod 3), k >= 4.