divrep oul: new option -fp<n> to force primes
Run with the option -fp<$n>, will explicitly set powers of each prime
up to and including $n.

This option must currently be given again when recovering, and it is
not expected to be safe to change the option between parts of a
recovered run.

Crude timings suggest that this will be a win for primes p such that
p^2 < k, but that may be less true if a high power of 2 divides n.