divrep oul: solve pell when we have two fixed squares
This should involve very few tests even with a high limit, so jump
straight into this when we spot it.

Expect a mess if we are unlimited.

This absorbs the Math::Pell module, which is intended to be released
separately to CPAN some time after a Math::Prime::Util release.