divrep oul: cope with overlarge -fp
We always set _something_ for a forced prime, so when the prime p exceeds
the number of elements we'll fail to test those values for which none of
the elements are divisible by p, eg with d == 1 (mod p).

Avoid a division by zero in that scenario, but also warn of it, and mark
this as not being a complete run.