divrep/oul: CHECKME bigint counters
The counters used in the square and non-square cases in walk_v are
initialized as plain ints, but used in contexts that are bounded
only by calculation time: probably enough to avoid loss of precision
in the non-square case, but not in the square case.

A brief check of the code suggests that, at least, it may be very
inefficiently converting these counters from raw integers to bigints
repeatedly for every calculation they are involved in.