divrep coul: check residues on new allocations when have_square
To do this efficiently, we pass an additional 'is_forced' parameter
to check_residue(): when not forced (and assuming vj != vi) we know
that p cannot divide vj - vi, and cannot divide v_{vj} and therefore m.

For now the original call, for when we first find a square, always
passes is_forced as true; we could save a little more by refining this.