divrep coul: gcc optimization
Use a set of optimization flags that have proven useful in the past.

This warns about ignored return values for ftruncate() in coul.c
and write() in diag.c, which should be addressed.

It also warns in many places of "missed loop optimization, the loop counter
may overflow [-Wunsafe-loop-optimizations]", which I don't fully understand;
I've turned this warning off for now, but not sure of the implications.

I _think_ it is warning that eg 'for (uint i = 1; i <= n; ++i)' can
overflow if n == UINT_MAX; if so, I need to check each of those and
perhaps add some asserts.

Ideal would be also to introduce '-fwhole-program', but I'll need to
experiment to see if I can make that practical.