divrep coul: allow higher debug levels
To look at cases, we set debug=1 to show every set of allocations;
but we show only the first iteration of walk_v, since there is no
useful new information to show for subsequent iterations. However
to debug and get timing stats on parallelization we _do_ want to
see each iteration, to separate additional diagnostics from one
case to the next; so for this we set debug>1.