divrep coul: verbose mode
Work in progress to get data on factorization timings and success.

Parallel testing has proven to be hugely faster in some cases - but
significantly slower in others. It is likely the tests should be
grouped into larger chunks, but it is unclear precisely how.

The changes here - which should have no net effect unless the code
is built with -DVERBOSE - attempt to capture timing and success for
each part of each factorization; it still needs further refinement
however before the data becomes useful.

What we hope to achieve is enough information to determine for a
given N in the region of 2^n: a) how long a give factorization test
takes if unsuccessful, as a function of n; b) how long on average
a given test takes to find a factor when successful, as a function
of n; and c) what the probability is that a given test is successful,
as a function of n. In each case the results are predicated on the
fact that all previous tests have been unsuccessful.

Even with that information, the decision does not become trivial:
the value of finding a singular factor of N_1 with required tau=4,
is much greater than that of finding a factor of N_2 with required
tau=8, for example - unless the chances of the remaining cofactor
being prime is high, in which case the benefit is reversed.