divrep coul: more rootmod bugfixes
1) _allrootmod_prime() has a sanity check that we're not finding more
roots that can exist - but it counted from the start of the results
array, which we're supposed to be appending to.

Clarify that the function appends, and fix the sanity check to count
from where we are instead of from zero.

2) _allrootmod_prime_power_r() is supposed to append results, but it
calls itself recursively.

Stash the existing results if there are any - which will only be at
the zeroth level of recursion - and restore them after the recursion
is done.