divrep coul: optional overflow check after update_chinese
When we find that current allocations require v_0 == rq (mod aq) with
rq > max, no solution with these allocations is possible with v_0 <= max.
However this case hits rarely, and saves little when it hits - the only
case that seemed to show a small overall improvement (about 0.02%) was
with capped searches for D(54,4), and that improvement is expected to
disappear if we improve the calculation of modular roots.

As such, include this check only if built with "-DCHECK_OVERFLOW".