divrep coul: bugfix power suppression
"improved power suppression" relied on stored vlevels for forced
primes to determine if the last allocation was a forced prime, but
the stored vlevels were not intended for that purpose, and did not
include data for the last forced prime. As such, we would wrongly
treat the last forced prime as unforced for the purposes of power
suppression, causing us for example to miss the solution for
D(54,3) because we suppressed allocations over 3^8.

Use reliable data (maxforce[]) to determine this instead.