divrep coul: defer "fits in ulong" check for walk_v
It is more convenient to let walk_v() walk values even if we will not
be able to walk the full range: we sometimes invoke this in every
confidence that a solution will be found well before the end of the
range.

So if the limit does not fit in ulong, set it temporarily to ULONG_MAX,
and emit the fatal error only if we manage to walk that far.