divrep coul: add last-resort factorization tests
These are invoked when built with -DPARALLEL and -DTRY_HARDER.

These are the 10 slowest sets of ECM tests from our linear factorization
code: the first set takes about 15 minutes to run (on my computer), each
subsequent set is about double the cost.

My observation is that if the preceding faster tests did not find
a factor, it is very rare that these tests do so: it is probably more
sensible to go find a GNFS solver at that point.