divrep coul: extend handling of outsized limit
If limit_p() returns 0 to signal that the limit is too large to fit
in ulong, any attempt to recurse at this level will fail. If we have
a forced square, we will attempt to handle that case by forcing
a call to walk_v() instead.

However, even without forced squares a walk_v() in this case may succeed;
and if it cannot succeed, it will give an appropriate error at that point.
So let it force a walk always.

This removes the diagnostic log line 002 "outsize limit without square".