divrep coul: bugfix tau_multi_prep
After a successful power check, must test if the result is prime.
Before this fix, in rare cases the code could abort with a message
like "no factors found for 2 non-primes starting 61 (9)".

Also an off-by-one testing the end condition for trial division
made it less efficient when the remnant was a small power.