coul: use a signal-based timer for progress diagnostics
This is in part an attempt to fix undiagnosed hangs on FreeBSD, which
may be due to timer wraparound.

However it should also let us spend less effort checking whether it is
time to show another progress update, at the risk of being less portable.

If the CLOCK_PROCESS_CPUTIME_ID clock is not available, we fall back
to CLOCK_REALTIME. In this case the initial "001" log line will show
"*RT*" at the end; the program will then be vulnerable to changes to
the system clock.