pcoul: new strategy -j2
Strategy 2 is the inverse of strategy 0: it prefers to allocate next
to the place with the _lowest_ remaining tau.

This strategy is not expected to be beneficial in most cases: it will
tend to result in an order of allocations that gives a small saving
in the walkc count and a moderate saving in the speed of walkc
iterations, but with a high cost of additional recursions.