inject: cope better with large numbers
If -l is used to increase the trial-factorization limit, apply that only
to v_i: i >= k. We use this to show a non-match by roughness, we don't
need it for matches.

Also apply the command-line hints before trying other factors, and
(in compensation) sort the factors found before printing.