coul: correctly recover stack that does not apply
When showing progress, we show the set of allocations just before
we attempt to apply the most recent one. So on recovery, the last
allocation to apply may turn out to be invalid, or have nothing
to do, or have so little to do that apply_single() chooses just
to do it.

In all those cases, in normal flow we would just shrug and go on
to try the next case. So on recovery we should do the same.

That does mean there is a risk that we fail to notice an actual
corrupt state. That is partially mitigated by the fact that we only
accept this for the last allocation we attempt to stack (and we
verify that all listed allocations were handled).