sq12: first pass
"sq12 <lim> <pmin> <pmax>" aims to rule out all possibilities for
a chain of 10 or more consecutive integers starting at n_0 <= lim,
such that at least one member of the chain is of the form p^2qr
for a prime p in the range pmin < p <= pmax.

If a counterexample is found, p^2qr is printed.