sq12: mod 576 check avoids some work
To check a value v, we first find n32 (the nearest multiple of 32), then
verify that it is of the form 32p, then check that n32 +/- 2 is of the
form 18q.

Speed that process up by first finding m = v (mod 576), then immediately
returning unless an array lookup tells us that m >> 4 is one of (9, 10,
25, 26). This automatically checks that n32 +/- 2 is divisible by 18,
and additionally that p is not divisible by 2 or 3, thus in many cases
saving an is_prime check and some additional calculations.