coul: new option -W<midp> for "two-stage booster"
Note: for now, -W is supported only in restricted cases for D(12,k),
but it may be extended in the future. We require that:
- n = 12;
- a maximum prime is set with -p;
- a single batch is specified with -b;
- the -W level midp is set such that 2 (midp + 1)^5 > max (set with -x),
  so that we won't skip any possible allocations of p^5.

Within those restrictions, if -W<midp> is speicified, then the batch
is processed in two stages: first, we try allocating p^2 at each possible
position in the pattern, for each midp < p <= maxp. For each allocation
we do a linear searh of possibilities. Second, we continue with normal
processing, but with the maximum prime set to midp.

During the first stage, we show "W(p, i)" in progress reports when
we have allocated the prime p at position i, and this is parsed on
recovery.