coul: display progress in walk_midp
If there is a fixed square, the apply_single/walk_v pair may go a long
time without finding a value to walk (that would trigger progress
display). So add a display also in the main walk_midp loop.

We need to jump through extra hoops, since the W() values we want to
show are not applied to value[] at the point we call diag_plain().