coul: simplify level handling for walk_midp()
If we define the "normal" case as that in which the level is already
correct, then it turns out that the only normal case is on recovery.
The abnormal case is when we are triggered from apply_batch(), when
(naturally enough) the current level is that for the batch we just
applied.