caul: handle force p^0 allocation
For addseq, when p divides n we may have a forced allocation of p^0,
to force the situation where _no_ element of the AP is divisible by p.
Since with the standard representation this would be represented by
a batch allocation { .vi = 0, .x = 0 }, we refine the representation
of a partially-forced p being unforced to { .vi = 1, .x = 0 }.

We must allow for that when checking for an unforced tail, and when
allocating the new case in apply_batch() or insert_stack().

(The new type of allocation is not yet generated.)