oul, coul: Pell bugfix
For the general Pell equation x^2 - Dy^2 = N, the algorithm implemented
finds coprime (x, y). For the full set of solutions, we must find each
k: k^2 | N and solve (kx)^2 - D(ky)^2 = N for coprime (x, y), interleaving
the results.

This was discovered trying to reproduce addseq(30,2) which tries to solve
162x^2 - 48y^2 = -30, which leads to x'^2 - 6y'^2 = -60. Since 6 is not
a quadratic residue of 60, this has no coprime solutions; however 6 _is_
a quadratic residue of 15, leading to solutions resulting in the
required x=43, y=79.

Must recheck all results obtained with this code. For oneseq(n,k) with
n divisible by only one odd prime, squares are never fixed by an unforced
prime, so we need recheck only patterns that fix 2 or more squares.
(We should force the maximum for safety.) For other n we need to recheck
more broadly, with SQONLY or something even more restrictive (skip walk
unless there are 2 squares fixed that differ by a multiple of a square).