bug recovery, Pell part 1 of 2
For the Pell bugfix, with n divisible by a single odd prime, we need
to rerun only those patterns that fix 2 or more squares. Since unforced
allocations never fix a square with such n, no other cases can reach
the Pell solver.

The recovery code parses 'report' output to find the list of confirmed
solutions, the runs 'pcoul -a' to find the batch id of each relevant
pattern, then runs each of those patterns. For each pattern we verify
that it actually completes (367 log) and echo any candidate it finds
(200 log); f(6,3) and f(6,4) verify that this works.

This has been run for the following n to show that no solutions were
missed:
  6, 10, 14, 22, 26, 34, 38, 46, 58, 62, 74, 82, 86, 94
  12, 20, 28, 44, 52, 68, 76, 92, 24, 40, 56, 88, 48, 80, 96

With a slight hack to allow it, it has also been run for D(12,12).
Rerunning D(24,11) with -f11 was not necessary since the original
run was with -f7.