bug recovery, Pell part 2 of 2
For the Pell bugfix, with n divisible by 2 or more odd primes, we need
to rerun all the Pell calculations.

We create a 'pellonly' executable which runs normally, but skips all
walk_* tests except those where two or more squares are fixed.

The recovery code parses 'report' output to find the list of confirmed
solutions. We then check logs to find the parameters used for the run,
and repeat it with pellonly.

For the cases f(66,4) and f(78,4) which were sharded by batch, we needed
to rerun the individual batches each with its own parameters.