caul: add TYPE_OFFSET to differentiate caul from coul
The difference between successive terms of the arithmetic progression
is always 1 for oneseq, but tau(n) for addseq. We use a new function
TYPE_OFFSET(diff) to represent the difference between two terms "diff"
apart.

When searching for related powers for a batch in apply_batch() or
insert_stack(), we check TYPE_OFFSET() for all offsets from the primary
rather than skipping along p at a time.