bug recovery, root_extract
root_extract() could return the wrong list of roots.

This is called when we upgrade from one fixed power to a higher fixed
power. For this to be reached, we need at least two distinct odd primes
p < q to divide n, and either:
- the pattern first fixes to a square, then fixes p to leave a q'th root;
- p > 3, and the pattern fixes a square.

For n <= 100, the second case only applies for n=70, we needed to rerun
SQONLY in full.

The first case applies for n in (30, 42, 60, 66, 78, 84, 90). For these
we needed to scan -a for relevant patterns, and rerun those patterns SQONLY.