coul: new option -R to skip recovery
This option can be useful to accumulate runs by individual batch:
we will append to the log file in the usual way, but will not attempt
to read it to find a recovery point.

Note that this also means we will not spot and truncate any corruption
at end of file following a reboot.

File management is rearranged: we open the file only for reading for
the purposes of recovery, and only for append for normal operation.