c?ul: experimental new strategy -j3
Intended specifically for cases with a significant squared factor.
For example with n=36, after allocating an unforced p^5 the only
remaining allocation possible at the same spot is another q^5 -
but that's not a square, so our residue filter does not kick in.

Under strategy j3, for such cases we always return the affected
location as the best. That is guaranteed to give a set of walk_1() calls,
which should quickly handle the case.