c?ul: add midp_recover.valid
Fix a bug on recovery: with -W but not W(...) in the progress line, the
manipulation of maxp and orig_maxp we performed to try to leave things
in the right state failed to take into account that we now allow more
than one batch.

Instead we now use a separate flag to determine whether the midp_recover
structure is valid, leaving maxp and orig_maxp alone.

We should now be able to rely on testing (maxp > midp) to indicate
that we need a walk_midp() run, (maxp == midp) implies we've already
done it for this batch.