c?ul: fix recovery in walk_midp on tail batch
During normal progress we call walk_midp from two main places - for a
full batch we do so within apply_batch(), but for a tail batch we do
so directly from recurse().

On entry to recurse() with jump_continue=IS_MIDP, signifying that
we were in the middle of a walk_midp(), we failed to distinguish
between those two sources. As a result, we did not continue correctly
after the walk_midp() call had completed.