c?ul.c: treat first forced power like others
When first looking at the set of possible forced powers of a prime,
we need some specific code: to check that the set is not empty, and
to stash the value levels ready for derecursing. That done, the
actual handling of the first forced power should be identical to
that for the rest - so use the same code for it.

In particular this ensures that we will spot it if the first entry
of a set indicates a tail batch (which currently it never does),
and opens up further refactoring opportunities.