c?ul: unify batch processing logic
There are two ways to reach the point of (knowing we have) a complete
batch: allocating each of the forced primes (a "full" batch), or
discovering that we've reached a requested tail on the current forced
prime (a "tail" batch).

At this point we need to decide whether to process the batch or skip it,
according to the batch options ("-b"). This is also the point we check
for large prime allocations if "-W" was specified, so in the future
will also be the point we wish to calibrate -W and -g if we want to
automate those.

So unify the logic into a new process_batch() function, with
concomitant simplifications.