c?ul: simplify/clarify midp handling
We call process_batch() only the first time we start a batch (and thus
in particular _not_ on recovery - we reenter walk_midp() in that case
by a different route). So in the general case, the test for whether we
should call walk_midp() is simply "if (midp)".

For diagnostics, we should consistently rely on in_midp to test whether
we are within a walk_midp() call.

We do not need to calculate maxp when none is provided: code is clearer
and more consistent if we simply handle "(orig_)maxp == 0" to signify
that no limit was specified.