c?ul: fix recovery from within walk_1_set()
The prep_unforced_x() call during stack insertion actually performs
the walk and returns PUX_SKIP_THIS_X in this case, so we can continue
correctly by recovering to the same point this would have reached.

The old comment "must pass in p=0 for it to calculate limp", it turns
out, was never true. Passing in the known p means we avoid redoing
the parts of the walk_1_set() call that we have already done.