c?ul: add support for -p100^2
For the minimum and maximum primes set with -p, and for the additional
threshold set with -W, we now detect and parse a value of the form
"100^2", and take it as a request to apply the relevant threshold to
the prime power p^e under consideration, instead of only to the prime p.

To enable this, we additionally now track "have_min" at each level to
say whether any allocation has fulfilled any minimum prime requirement.