c?ul: don't pass level to disp_batch()
It already calls prep_show_v() which itself assumes levels[level], so
it is pointless to let disp_batch() look at anything different.

The underlying intent was for both disp_batch() and prep_show_v() to
show data related to a specified level; however that isn't currently
(reasonably) possible, since when we show the core values[] data we
don't know which level it comes from. If that changes at some point,
this should be revisited.