c?ul: fix level for batches
It turns out that until 9dce3ed935, passing a level to disp_batch() was
masking the fact that the level was wrong (probably since 6feff02875 or
before), which could have introduced subtle errors.

walk_midp() adjusted to account for that, so we must now bump the level
around that call.

This is an ongoing mess; to fix it, the plan is to move tracking of
values[vi].vlevel into the levels[] structure and pass a cur_level
pointer to anything that needs it. Eventually, the only things looking
at 'level' directly should be those that are changing it, primarily
recurse() itself.