gtauseq: propagate constraints on init
When we first initialize the constraint vector for some modulus m by
combining existing vectors m/p for each p dividing m, we must check
if the resulting vector shows new results that can be propagated back
to its divisors.

This was noticed from an apparent hang of track(123, 8) since track(123, 7)
had a fixed power, but due to the lack of propagation (coupled with a
slightly different order of applying constraints) track(123, 8) did not
detect that power.

Note that this could cause suboptimal behaviour by underconstraining,
but not erroneous behaviour.