pc?ul: copy/paste bug in walk_1_set
Values fixed to be squares were tested as primes instead.

This may require some calculations to be redone; however few if any
cases are expected, since in most cases if any values are fixed to be
square we don't do the further allocations that could cause this path
to be reached.