pcrul: cater for initial squares
If initial set of target_t indicates any squares, reserve level[1] as
a dummy level in which to apply alloc_square(). If two or more squares
are indicated, go immediately to walk_v() for a Pell-based solution
without ever calling recurse().

FIXME: this increments 'final_level' to indicate an extra level through
which we should not attempt to derecurse; however we now need a separate
way to determine which forced prime should be allocated at which level,
robust both for this case and for the 'init_pattern' case.