harness: new run-type shard-test (Seq::Run::ShardTest)
From time to time, as runs get slower, we run the next shard test for
a given f(n,k). For the m'th shard test, we do an initial run to find
the current fixed modulus M; we then try new runs forced in turn to
each possible value (mod Mm) and track which of those finds a
contradiction in the constraints. Those are then captured as newly
disallowed moduli in tauf.optm.

When an f(n,k-1) solution is found, we propagate the optm values for
f(n,k) but reset the record of completed shard tests to 0, so that
the additional constraint of a longer sequence has the opportunity
to find additionally disallowed moduli.