harness: fix tracking of last range checked
To pick appropriate values for the next gtauseq run, we want to know how
long the last relevant run took and what range of numbers it checked.
Since the range of numbers checked may be both smaller than that requested
(if a solution was found) or larger (if some - maybe many - numbers were
skipped by modular checks beyond the end of the requested range), we cannot
rely on the recorded "optx" value for the run; however if another process
such as upperlim has been used in the meantime, we also cannot rely on
the "checked" limit to tell us where _our_ last run reached.

So a) on finalizing a run, update optx to record the actual maximum value
checked; b) ensure the database entry for the run is updated before we
ask what the next run should like like; c) take greater care to ensure
that the "last (relevant) run" looked at was of the right type (ie that
it was one with owner "harness", not owner "upperlim").

Now simplify Seq::TauF::_strategy to use optx as the upper end of the
last range.

TODO: if no relevant last run is found for the current k, we ask Seq::Run
for the last run of a previous k. However this function does not strip
out optimizing runs. If such a run is returned, it may appear that this
case runs slower than is actually the case, affecting various factors
about the next run we will attempt.