pc?ul: new option -dl to show prefinal results for harness
With this option, we additionally write some log lines that are relied
on by the run harness: on failure, we write a 500 line showing how far
we tested (and timing); on success we write a series of 211 lines showing
the actual factorization of the best candidate and its successors.

TODO: replace write_fact(), which performs a full factorization, with
something that has more of the logic of 'inject' to avoid unnecessary
factorization effort.