pc?ul: choose order to test Pell solutions
Simplistic: choose the one with the lowest tau, or if equal, the one
with the highest divisor (and thus the lowest value) to test first.
This can become significant in cases where one or more Pell equations
give the _only_ possible solution(s) (such as track(324,3) for example),
since we may then quickly find ourselves testing very large numbers.