Type::Track: suppress non-residues on fixed powers (disabled)
When we discover v_k = xy^z, the main apply() checks will only have
the opportunity to notice that when x = 1 (or x square).

Add code to suppress non-residues for the x > 1 cases; but leave it
disabled, since the effects for the main fixed power will in any case
be implicitly handled when we fix it, so this adds something only for
the discarded secondary fixed power(s) in the Pell case.