c?ul: also handle squares in parallel
Squares are not guaranteed to be "too small to care about" - rather than
block factorizing those before handling other non-primes in parallel,
include them in the list to parallelize.

We still do the tau_multi_prep() call early, so the simple cases can
still short-circuit.

This can now complete testing of the track(324,3) Pell solution near
1e931 instantly (since v_2 = 3^2.C930 but requires tau=4), instead
of stalling for many hours trying to factorize the square root of
v_0 (tau=15).