gtauseq: replace mod_combine
When mod_combine() fails, we want to log (rather than die) with the
error - it is a perfectly normal case of "no remaining valid values".
So replace it with a simpler implementation that uses chinese() for
the bulk of the work, and returns (0, 0) for the failure case.

Along the way, cleanup other unused ModFunc:: functions, and unused
imports.