c?ul: detect more of the impossible
If no valid batch was detected, log this in the same way as
"no arrangement of powers [of a fixed prime] is valid".

We need to take care in the walk_1() case that we treat it as a
"nothing valid" case if the fixed value implies a base less than min.