harness: prioritize recalibration after finding result
After finding a result for f(n,k), we use the details of that run to
decide the parameters for the first run of f(n,k+1). However the
timings of that run are less relevant (since it was not a complete
run), so artificially increase the priority of the next run to ensure
we get relevant f(n,k+1) timings as soon as possible.