TRACK_STATS: fix stats for tau_multi_run()
Modify the return value of tau_multi_run() to be the number of remaining
numbers to test (so that 0 means success) instead of a boolean success,
so we can track stats.

Note that a significant proportion of values may be rejected in the
tau_multi_prep() stage, which will skew the results.