pc?ul: new option -P<n> to control recurse/walk switch
This option specifies a cap on the effective maximum prime used in
the calculation that determines whether to recurse further or walk
what we have. It is independent of gain (-g).

The intention is to allow, for example, a pair of runs with -p50 and
-p50:100 to test all the same cases that a single -p100 run would test.
If both are run with -P50 and other options are kept the same through
all runs, I *think* they should be guaranteed to do so.

The -P value for all runs should be no larger than the least -p value
in any run, ie 50 in the example above.