pc?ul: -F<n> is buggy if n too low
We handle unforced primes with a "tail" entry which switches us from
forced to unforced mode in the recursion; however there may still be
higher primes that should have forced patterns applied. Thus
'pcoul -F3 160 6' will miss a pattern like "5^4 2 . 2^4 . 2.5"
due to the tail handling for 3^x.

For now we check for and disallow this case, until we can work out
how to handle it correctly.

Checking logs, the only runs needing to be corrected were the two that
alerted me to the issue.