pc?ul: support -o
On reach the end of the selected factorization strategies without
success, log the current target and each of the numbers that are still
to be factorized (along with their source v_i and target tau).

When -o is not enabled, similar info will be shown on the console but
not logged.

TODO: add option to stop factorization earlier, perhaps as "-o65"
to fail before attempting QS (or (most) later strategies) on numbers
of 65+ digits. It should be possible to achieve this by adapting
the coultau.c:tmfb[] array (though easier if the option takes a bit
size rather than a decimal size).