find-2pqr <n>
Given n of the form 2pqr, {p, q, r} in P >= 3, this finds the maximum
possible value of v_0 and the minimum possible value of v_1 to check
if we can immediately rule out g(n) >= 6.

For n < 1000 this rules out all but:
  350 490 510 550 570 650 686 690 714 726 770 798 850 858 870 910 930 950 966
.. and shows ranges that would need explicit checks to rule out those
additional cases.

It also finds g(162) < 6, since we cannot split that into 5 factors of which
one satisfies the need for 2^4.