pcoul: special-case strategy for n == 2 (mod 4)
When n == 2 (mod 4), if a chain has values n_0 == 0 (mod 8) and
n_2 == 2 (mod 8) then we require n_2 = 2x^2 => n_0 = 2(x-1)(x+1).
This is a powerful constraint, so we add a new strategy to kick in
whenever we spot that it can apply.

We check if it can apply during apply_secondary(), which is very low
cost. When it does apply, we change to a new strategy STRATEGY_6X
which a) checks if it still applies (and restores the original
strategy if not); b) checks if the n_0 target now wants only a prime
(and asks for more allocations to the target if not), then c) checks
each of the possible values for x directly to complete the processing.

For case (b) we also add a new check in limit_p(), since additional
allocations can be restricted to a much smaller range in this case.