pc?ul: improve ability to find "no valid arrangement"
Singletons set seen_valid only if they find a solution; new option
'-jp' for validity testing - by default we make an early call to
walk_v() as soon as we see two values forced to be squares, to take
advantage of the resulting Pell equation; when '-jp' ("defer_pell")
is in operation, we don't call walk_v() until we have a full batch.
(As a result this also honours the usual constraints of -a and -b
options to avoid walking at all.)

Commit ad7c7cf834 introduced the early Pell handling, and took
advantage of it to remove the code in the Pell branch of walk_v()
that catered for the possibility there were more than 2 square values
to test. So now we have to restore that code.