pc?ul: new strategy -j4 to allow allocation of 2^x'th powers
This is expected to be useful for cases where tau is divisible by
a high power of 2.

New strategy '-j4' acts like -j1 for powers with odd divisors, and aims
to minimize the range of the next prime for 2^x'th powers. We order
things such that an allocation of p^{2^x} assumes what remains is not
divisible by any prime <= p.

When this strategy is chosen, we set the 'highpow' flag'. When it is
set, we maintain an additional mpz_t t_allocation.lim with every
allocation as the maximum roughness of what remains; the strategy uses
this to prefer the target with the smallest range.

'-p' and '-W' also apply limits to 2^x'th powers when highpow is set.

TODO: consider whether there is something significant to be gained in
walk by taking advantage of "we now assume v_i/q_i is divisible by no
prime less than that of the last allocation", either/both in early
rejection (cf inv[]) or in factorization. A nice side-effect of early
rejection would be to reduce duplicate candidates.