pc?ul: accurate cached mintau()
We calculate exact mintau() recursively, and cache the result.
This is currently marginally slower for small tau (<= 100), but already
significantly faster for larger tau. The caching should also ensure that
the longer the run the more time should be gained compared to the previous
implementation.

Further improvements should probably be enough to match or beat the
previous implementation even for small tau:
- keep long-lived arrays of mpz_t mint_best and mint_px indexed by depth0,
to reduce malloc churn;
- maintain the bit vector of available primes in levels[] rather than
rebuilding from scratch on each call;
- optionally, short-circuit for prime tau.