pc?ul: better mintau with repeated high factors
Let h(x) be the highest prime factor of x. When assigning p^x with
remaining tau t such that h(x) = h(t/x), we know that future allocations
1) must avoid powers q^y: h(y) = h(x), y < x; and 2) that a future
allocation q^x must have q > p.

To take advantage of this, we now have a parallel version of mintau(),
mintau_restricted(), which takes that x as an additional parameter, and
finds (and caches) mintau values that avoid y <= x. In limit_p(), we
use this in combination with as many copies of q^x as are possible
to find an appropriately restrictive limit on p - which may also find
that no valid p remains.

We maintain a parallel cache for each possible restriction.