pc?ul: new strategy -js0,1,2
New strategy -js takes a comma-separated list of offsets, and asserts
that these offsets should be chosen, in this order, for every allocation.
Reaching the end of the list forces remaining values to be walked.

A given order is unlikely to be sensible for multiple batches, so this
will usually be used with -b (or such -m as are enough to fix the batch).

TODO: add option to fall back to some other strategy at the end of the list.
FIXME: we would have liked to set highpow dynamically, based on remaining
tau of the next v_i about to be returned; but we rely on it being set at
various points during setup. One option is to set it mandatorily when we
see -js; another, more useful option would be to make it an additional
strategy option (-jh?) at which point -j4 gets absorbed to -j0 -jh (and
other combinations automatically become available, for better or worse).