pc?ul: make -I handling a bit saner
1) commit fbf3ac3d added a mandatory fail() on any use of -I with the
message "fix the FIXME for chinese() combine first". Since the referenced
fixme is reached only when some check or modfix option leaves us with
a global modular constraint, we can at least reduce the fail() to occur
only if one of those options is also used.

2) it is common to recover an interrupted run with the same options as
when it was initially launched, so presence of -I should not act as if
-R has also been specified - the user should add -R if that is the
behaviour they want.