pc?ul: extended form for -a
The option '-a' now takes an optional numeric argument (default 1).
If bit-value 1 is set, as before, we list all batches, skipping most
processing; if bit-value 2 is set, we provide extended output of the
batch information in CSV form and skip all other output; if bit-value
4 is set we skip even the processing that is normally done in passing,
eg for patterns where a value is fully allocated, or where multiple
squares result in a Pell equation.